+++ src/modules/tree-store/store/use-dataset-tree-store.ts
@@ -5,9 +5,12 @@ import TreeStore from "src/modules/tree-store/utils/tree-store/tree-store";
 import treeStoreStorage from "src/modules/tree-store/storage/tree-store-storage";
 
 import type { INode, NodeIdType } from "src/modules/tree-store/utils/tree-store/types";
+import useAppEnvironment from "src/composable/use-app-environment";
 
 const useDatasetTreeStore = defineStore("dataset-tree-store", () => {
   const treeStoresMap = reactive<Map<string, TreeStore>>(new Map());
+
+  const { canSaveData } = useAppEnvironment();
 
   const loadAll = async () => {
     const resData = await treeStoreStorage.getAll();
@@ -21,13 +24,17 @@ const useDatasetTreeStore = defineStore("dataset-tree-store", () => {
   };
 
   const saveTreeStore = async (name: string, data: INode[]): Promise<void> => {
-    await treeStoreStorage.save({ name, data });
+    if (canSaveData.value) {
+      await treeStoreStorage.save({ name, data });
+    }
 
     treeStoresMap.set(name, new TreeStore(data));
   };
 
   const deleteTreeStore = async (name: string): Promise<void> => {
-    await treeStoreStorage.delete(name);
+    if (canSaveData.value) {
+      await treeStoreStorage.delete(name);
+    }
 
     treeStoresMap.delete(name);
     if (selectedDatasetName.value === name) {
@@ -50,10 +57,12 @@ const useDatasetTreeStore = defineStore("dataset-tree-store", () => {
   const persistSelectedTreeStore = async () => {
     if (!selectedTreeStore.value) return;
 
-    await treeStoreStorage.save({
-      name: selectedDatasetName.value,
-      data: selectedTreeStore.value.getAll().map(node => toRaw(node)),
-    });
+    if (canSaveData.value) {
+      await treeStoreStorage.save({
+        name: selectedDatasetName.value,
+        data: selectedTreeStore.value.getAll().map(node => toRaw(node)),
+      });
+    }
   };
 
   const addNode = async (newNode: INode): Promise<void> => {
